+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Reload configuration files when their symbolic link is replaced
 - Strip trailing whitespaces when yanking from a block selection
 - Display area keeps history position when viewport is cleared
+- Commands spawn from the current directory of the foreground shell in Unix-like systems
 
 ### Fixed